+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Changelog
 - Fix crash when the file is not uploaded using ``attachment`` field name (fixes #57)
 - Prevent crash when filename is not provided (fixes #81)
 - Update the call to the Record resource to use named attributes. (#97)
+- Show detailed error when data is not posted with multipart content-type.
 
 **Internal changes**